requirements.txt
python-telegram-bot[job-queue]==21.6
pandas==2.2.3
openpyxl==3.1.5
ping3==4.0.8
python-dotenv==1.0.1
